cleaner short address handling
